100-get_endianness.c fixed

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -8,7 +8,7 @@
  */
 int get_endianness(void)
 {
-	short int m = 1;
+	unsigned int m = 1;
 	char *c = (char *)&m;
 
 	if (*c)
